sd-backend/requirements.txt
torch
diffusers
transformers
accelerate
fastapi
uvicorn
pillow

